Remove useless class and add get method to ModModButtonController
